+++ src/core/getFrameValidatedMessage.test.ts
@@ -1,18 +1,10 @@
 import { mockNeynarResponse } from './mock';
 import { getFrameValidatedMessage } from './getFrameValidatedMessage';
+import {neynarBulkUserLookup} from "../utils/neynar/user/neynarUserFunctions";
 
-const bulkUserLookupMock = jest.fn();
-jest.mock('../internal/neynar/neynarClient', () => {
+jest.mock('../utils/neynar/user/neynarUserFunctions', () => {
   return {
-    NeynarClient: jest.fn().mockImplementation(() => {
-      return {
-        user: {
-          bulkUserLookup: bulkUserLookupMock,
-          // other user functions can be mocked here
-        },
-        // other properties and methods of NeynarClient can be mocked here
-      };
-    }),
+    neynarBulkUserLookup: jest.fn()
   };
 });
 
@@ -31,7 +23,7 @@ describe('getFrameValidatedMessage', () => {
   it('should return undefined if the message is invalid', async () => {
     const fid = 1234;
     const addresses = ['0xaddr1'];
-    const { validateMock } = mockNeynarResponse(fid, addresses, bulkUserLookupMock);
+    const { validateMock } = mockNeynarResponse(fid, addresses, neynarBulkUserLookup as jest.Mock);
     validateMock.mockClear();
     validateMock.mockResolvedValue({
       isOk: () => {
@@ -47,7 +39,7 @@ describe('getFrameValidatedMessage', () => {
   it('should return the message if the message is valid', async () => {
     const fid = 1234;
     const addresses = ['0xaddr1'];
-    mockNeynarResponse(fid, addresses, bulkUserLookupMock);
+    mockNeynarResponse(fid, addresses, neynarBulkUserLookup as jest.Mock);
     const fakeFrameData = {
       trustedData: {},
     };
