+++ src/utils/neynar/exceptions/FetchError.ts
