+++ src/utils/neynar/user/neynarUserFunctions.test.ts
@@ -1,5 +1,6 @@
-import { NeynarClient } from '../neynarClient';
 import { FetchError } from '../exceptions/FetchError';
+import {neynarBulkUserLookup} from "./neynarUserFunctions";
+
 
 describe('neynar user functions', () => {
   let fetchMock = jest.fn();
@@ -19,15 +20,14 @@ describe('neynar user functions', () => {
     fetchMock.mockResolvedValue({
       users: [{ fid: 1 }],
     });
-    const client = new NeynarClient();
-    const resp = await client.user.bulkUserLookup([1]);
+
+    const resp = await neynarBulkUserLookup([1]);
     expect(resp?.users[0]?.fid).toEqual(1);
   });
 
   it('fails on a non-200', async () => {
     status = 401;
-    const client = new NeynarClient();
-    const resp = client.user.bulkUserLookup([1]);
+    const resp = neynarBulkUserLookup([1]);
     await expect(resp).rejects.toThrow(FetchError);
   });
 });
